requirements.txt
flwr
torch
torchvision
torchaudio
numpy
scikit-learn
phe
pyyaml
psutil
requests
protobuf
grpcio